Fix cs/loss-of-precision CodeQL alerts (batch 30): int-mul overflow before double cast (#1568)
## Summary

Fixes 4 open CodeQL `cs/loss-of-precision` alerts (batch 30 of the
ongoing Tier 4 quality-rule campaign tracked in #1380). All 4 share the
same root cause: "Possible overflow: result of integer multiplication
cast to double" - an `int * int` multiplication is computed in 32-bit
space, and only the *result* is widened to `double`, so the
multiplication itself can silently overflow (wrap) before ever reaching
the double.

**Fix approach:** widen one operand to `double` *before* the
multiplication (not after), so the multiply itself happens in double
space and cannot overflow as 32-bit int math. Added an explanatory
comment at each site. No behavior change for realistic inputs; only
removes a theoretical silent-overflow path for very large parsed/stored
values.

## Alerts fixed
- #718 / #717 -
`src/orcaslicer-worker/Services/OrcaSlicingPipelineService.cs:1257`
(PrintTimeSeconds computation)
- #716 - `src/infra/Services/JobSchedulingService.cs:811`
(`AdvanceSchedule`, Weekly recurrence branch)
- #715 - `src/infra/Services/Gcode/GcodeMetadataExtractorService.cs:632`
(`TryParsePrintTime`)

## Regression tests added
-
`GcodeMetadataExtractorServiceTests.ExtractMetadataAsync_HandlesExtremelyLargePrintTimeWithoutIntOverflow`
- pins the fix with a 40,000,000-hour print time (40M * 60 = 2.4B,
exceeds int.MaxValue).
-
`SchedulerOccurrenceSemanticsTests.TriggerScheduledJobs_WeeklyRecurrenceWithOverflowingInterval_DoesNotSilentlyCorruptScheduledDate`
- pins the fix with a RecurrenceInterval (613,566,756) chosen so the
pre-fix int multiplication wraps to `-4`, which would have silently
moved the schedule *backward* with no error. Post-fix, the true
out-of-range magnitude correctly surfaces as a caught/logged
per-schedule failure (schedule's `ScheduledStartTime` left unchanged)
instead of silent data corruption.

## Validation
- `dotnet build ./farm-web.sln -c Debug` - 0 warnings, 0 errors.
- `dotnet test ./farm-web.sln -c Debug --filter
"FullyQualifiedName~GcodeMetadataExtractorServiceTests|FullyQualifiedName~SchedulerOccurrenceSemanticsTests"`
- 60/60 passed.
- Branch synced with `origin/development` before opening this PR (merge
commit `e791795ea`, clean, no conflicts) and re-validated on the merged
result.

## Review
Reviewed by Bishop, Hicks, and Vasquez per the repo's pre-PR 3-way
adversarial review gate. Consensus APPROVE (self-attested per
`.github/copilot-instructions.md` § "Repository verdict evidence" - see
PR comments for canonical
`Squad-Reviewer`/`Squad-Verdict`/`Squad-Head-SHA` records).

## Issue linkage
Part of #1380 (Tier 4 CodeQL quality-rule campaign tracker - stays open
for future batches). Alerts #715, #716, #717, #718 are CodeQL alert
numbers, not GitHub issues, so they are not referenced with `Closes #`;
they will be dismissed directly via the code-scanning API once this PR
merges.

---------

Copilot-Session: 39577b36-cbbd-4ee1-8a97-e85996c81141

+++ src/infra/Services/Gcode/GcodeMetadataExtractorService.cs
@@ -629,7 +629,10 @@ public class GcodeMetadataExtractorService(ILogger<GcodeMetadataExtractorService
             int h = int.Parse(hoursMatch.Groups[1].Value);
             int m = int.Parse(hoursMatch.Groups[2].Value);
             int s = int.Parse(hoursMatch.Groups[3].Value);
-            minutes = (h * 60) + m + (s / 60.0);
+
+            // Multiply in double space so a large parsed hour value cannot overflow as int arithmetic
+            // before it is added into the double `minutes` result.
+            minutes = ((double)h * 60) + m + (s / 60.0);
             return true;
         }
